requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.3
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pandas==2.1.4
tqdm==4.66.1
openai==1.13.3
jinja2==3.1.3
pydantic==2.6.1
email-validator==2.1.0
xlsxwriter==3.1.9 